Batch mode for install

+++ lib/util.go
@@ -6,12 +6,14 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"regexp"
-	"strings"
 
 	"github.com/marwanhawari/stew/constants"
+	"github.com/marwanhawari/stew/lib/config"
+	"github.com/marwanhawari/stew/lib/ui/progress"
+	"github.com/marwanhawari/stew/lib/ui/prompt"
+	"github.com/marwanhawari/stew/lib/ui/terminal"
 	"github.com/mholt/archiver"
-	progressbar "github.com/schollz/progressbar/v3"
+	"github.com/pkg/errors"
 )
 
 func isArchiveFile(filePath string) bool {
@@ -25,46 +27,24 @@ func isArchiveFile(filePath string) bool {
 func isExecutableFile(filePath string) (bool, error) {
 	fileInfo, err := os.Stat(filePath)
 	if err != nil {
-		return false, err
+		return false, errors.WithStack(err)
 	}
 
 	filePerm := fileInfo.Mode()
-	isExecutable := filePerm&0111 != 0
+	isExecutable := filePerm&0o111 != 0
 
 	return isExecutable, nil
 }
 
-// CatchAndExit will catch errors and immediately exit
-func CatchAndExit(err error) {
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-}
-
-// PathExists checks if a given path exists
-func PathExists(path string) (bool, error) {
-	_, err := os.Stat(path)
-
-	if err != nil {
-		if os.IsNotExist(err) {
-			return false, nil
-		}
-		return false, err
-	}
-
-	return true, nil
-}
-
-// DownloadFile will download a file from url to a given path
-func DownloadFile(downloadPath string, url string) error {
-	sp := constants.LoadingSpinner
+// DownloadFile will download a file from url to a given path.
+func DownloadFile(rt config.Runtime, downloadPath string, url string) error {
+	sp := progress.Spinner(rt)
 	sp.Start()
 	resp, err := http.Get(url)
 	sp.Stop()
 
 	if err != nil {
-		return err
+		return errors.WithStack(err)
 	}
 
 	defer resp.Body.Close()
@@ -75,56 +55,49 @@ func DownloadFile(downloadPath string, url string) error {
 
 	outputFile, err := os.Create(downloadPath)
 	if err != nil {
-		return err
+		return errors.WithStack(err)
 	}
 	defer outputFile.Close()
 
-	bar := progressbar.DefaultBytes(
-		resp.ContentLength,
-		"⬇️  Downloading asset:",
+	bar := progress.Bar(
+		rt, resp.ContentLength, "⬇️  Downloading asset:",
 	)
 	_, err = io.Copy(io.MultiWriter(outputFile, bar), resp.Body)
 	if err != nil {
-		return err
-	}
-
-	_, err = io.Copy(outputFile, resp.Body)
-	if err != nil {
-		return err
+		return errors.WithStack(err)
 	}
 
 	return nil
-
 }
 
 func copyFile(srcFile, destFile string) error {
 	srcContents, err := os.Open(srcFile)
 	if err != nil {
-		return err
+		return errors.WithStack(err)
 	}
 	defer srcContents.Close()
 
 	destContents, err := os.Create(destFile)
 	if err != nil {
-		return err
+		return errors.WithStack(err)
 	}
 	defer destContents.Close()
 
 	_, err = io.Copy(destContents, srcContents)
 	if err != nil {
-		return err
+		return errors.WithStack(err)
 	}
 
-	err = os.Chmod(destFile, 0755)
+	err = os.Chmod(destFile, 0o755)
 	if err != nil {
-		return err
+		return errors.WithStack(err)
 	}
 
 	return nil
 }
 
 func walkDir(rootDir string) ([]string, error) {
-	allFilePaths := []string{}
+	var allFilePaths []string
 	err := filepath.Walk(rootDir, func(filePath string, fileInfo os.FileInfo, err error) error {
 		if !fileInfo.IsDir() {
 			allFilePaths = append(allFilePaths, filePath)
@@ -134,20 +107,19 @@ func walkDir(rootDir string) ([]string, error) {
 	return allFilePaths, err
 }
 
-func getBinary(filePaths []string, repo string) (string, string, error) {
+func getBinary(io terminal.Terminal, filePaths []string, installation Installation) (string, string, error) {
 	binaryFile := ""
 	binaryName := ""
-	var err error
-	executableFiles := []string{}
+	var executableFiles []string
 	for _, fullPath := range filePaths {
 		fileNameBase := filepath.Base(fullPath)
 		fileIsExecutable, err := isExecutableFile(fullPath)
 		if err != nil {
 			return "", "", err
 		}
-		if fileNameBase == repo && fileIsExecutable {
+		if fileNameBase == installation.Repo && fileIsExecutable {
 			binaryFile = fullPath
-			binaryName = repo
+			binaryName = installation.Repo
 			executableFiles = append(executableFiles, fullPath)
 		} else if filepath.Ext(fullPath) == ".exe" {
 			binaryFile = fullPath
@@ -163,22 +135,34 @@ func getBinary(filePaths []string, repo string) (string, string, error) {
 			binaryFile = executableFiles[0]
 			binaryName = filepath.Base(binaryFile)
 		} else if len(executableFiles) != 1 {
-			binaryFile, err = WarningPromptSelect("Could not automatically detect the binary. Please select it manually:", filePaths)
+			if installation.BatchMode {
+				return "", "", CouldntDetectBinaryError{filePaths}
+			}
+			var err error
+			binaryFile, err = prompt.SelectWarn(
+				io, "Could not automatically detect the binary. "+
+					"Please select it manually:", filePaths)
 			if err != nil {
 				return "", "", err
 			}
 			binaryName = filepath.Base(binaryFile)
-			binaryName, err = PromptRenameBinary(binaryName)
+			binaryName, err = PromptRenameBinary(io, RenameBinaryArgs{
+				Default: binaryName,
+			})
 			if err != nil {
-				return "", "", nil
+				return "", "", err
 			}
 		}
+	}
+
+	if installation.BinaryName != "" {
+		binaryName = installation.BinaryName
 	}
 
 	return binaryFile, binaryName, nil
 }
 
-// ValidateCLIInput makes sure the CLI input isn't empty
+// ValidateCLIInput makes sure the CLI input isn't empty.
 func ValidateCLIInput(cliInput string) error {
 	if cliInput == "" {
 		return EmptyCLIInputError{}
@@ -187,7 +171,7 @@ func ValidateCLIInput(cliInput string) error {
 	return nil
 }
 
-// CLIInput contains information about the parsed CLI input
+// CLIInput contains information about the parsed CLI input.
 type CLIInput struct {
 	IsGithubInput bool
 	Owner         string
@@ -195,68 +179,66 @@ type CLIInput struct {
 	Tag           string
 	Asset         string
 	DownloadURL   string
+	BinaryName    string
 }
 
-// ParseCLIInput creates a new instance of the CLIInput struct
+// ParseCLIInput creates a new instance of the CLIInput struct.
 func ParseCLIInput(cliInput string) (CLIInput, error) {
 	err := ValidateCLIInput(cliInput)
 	if err != nil {
 		return CLIInput{}, err
 	}
 
-	reGithub, err := regexp.Compile(constants.RegexGithub)
-	if err != nil {
-		return CLIInput{}, err
-	}
-	reURL, err := regexp.Compile(constants.RegexURL)
-	if err != nil {
-		return CLIInput{}, err
-	}
 	var parsedInput CLIInput
-	if reGithub.MatchString(cliInput) {
+	switch {
+	case constants.RegexGithub.MatchString(cliInput):
 		parsedInput, err = parseGithubInput(cliInput)
-	} else if reURL.MatchString(cliInput) {
+	case constants.RegexURL.MatchString(cliInput):
 		parsedInput, err = parseURLInput(cliInput)
-	} else {
-		return CLIInput{}, UnrecognizedInputError{}
+	default:
+		return CLIInput{}, UnrecognizedInputError{
+			Input: cliInput,
+		}
 	}
 	if err != nil {
 		return CLIInput{}, err
 	}
 
 	return parsedInput, nil
-
 }
 
 func parseGithubInput(cliInput string) (CLIInput, error) {
-	parsedInput := CLIInput{}
-	parsedInput.IsGithubInput = true
-	trimmedString := strings.Trim(strings.Trim(strings.Trim(strings.TrimSpace(cliInput), "/"), "@"), "::")
-	splitInput := strings.SplitN(trimmedString, "@", 2)
-
-	ownerAndRepo := splitInput[0]
-	splitOwnerAndRepo := strings.SplitN(ownerAndRepo, "/", 2)
-	parsedInput.Owner = splitOwnerAndRepo[0]
-	parsedInput.Repo = splitOwnerAndRepo[1]
-
-	if len(splitInput) == 2 {
-		tagAndAsset := splitInput[1]
-		splitTagAndAsset := strings.SplitN(tagAndAsset, "::", 2)
-		parsedInput.Tag = splitTagAndAsset[0]
-		if len(splitTagAndAsset) == 2 {
-			parsedInput.Asset = splitTagAndAsset[1]
+	parsedInput := CLIInput{
+		IsGithubInput: true,
+	}
+	m := constants.RegexGithub.FindStringSubmatch(cliInput)
+	if len(m) != 6 {
+		return CLIInput{}, UnrecognizedInputError{
+			Input: cliInput,
 		}
 	}
+	parsedInput.Owner = m[1]
+	parsedInput.Repo = m[2]
+	parsedInput.Tag = m[3]
+	parsedInput.Asset = m[4]
+	parsedInput.BinaryName = m[5]
 
 	return parsedInput, nil
-
 }
 
 func parseURLInput(cliInput string) (CLIInput, error) {
-	return CLIInput{IsGithubInput: false, Asset: filepath.Base(cliInput), DownloadURL: cliInput}, nil
+	if !constants.RegexURL.MatchString(cliInput) {
+		return CLIInput{}, UnrecognizedInputError{
+			Input: cliInput,
+		}
+	}
+	return CLIInput{
+		Asset:       filepath.Base(cliInput),
+		DownloadURL: cliInput,
+	}, nil
 }
 
-// Contains checks if a string slice contains a given target
+// Contains checks if a string slice contains a given target.
 func Contains(slice []string, target string) (int, bool) {
 	for index, element := range slice {
 		if target == element {
@@ -266,37 +248,58 @@ func Contains(slice []string, target string) (int, bool) {
 	return -1, false
 }
 
-func extractBinary(downloadedFilePath, tmpExtractionPath string) error {
-	isArchive := isArchiveFile(downloadedFilePath)
-	if isArchive {
-		err := archiver.Unarchive(downloadedFilePath, tmpExtractionPath)
-		if err != nil {
-			return err
-		}
+// Installation contains information about the installation process.
+type Installation struct {
+	DownloadedFilePath string
+	Repo               string
+	BinaryName         string
+	BatchMode          bool
+}
 
+func extractBinary(io terminal.Terminal, installation Installation, tmpExtractionPath string) error {
+	if isArchiveFile(installation.DownloadedFilePath) {
+		err := archiver.Unarchive(installation.DownloadedFilePath, tmpExtractionPath)
+		if err != nil {
+			return errors.WithStack(err)
+		}
+		return nil
+	}
+	renameArgs := RenameBinaryArgs{
+		Default:   installation.BinaryName,
+		BatchMode: installation.BatchMode,
+	}
+	if renameArgs.Default == "" {
+		renameArgs.Default = filepath.Base(installation.DownloadedFilePath)
 	} else {
-		originalBinaryName := filepath.Base(downloadedFilePath)
+		// If the user specified a binary name, we assume they want to use it
+		renameArgs.BatchMode = true
+	}
 
-		renamedBinaryName, err := PromptRenameBinary(originalBinaryName)
-		if err != nil {
-			return err
-		}
-		err = copyFile(downloadedFilePath, filepath.Join(tmpExtractionPath, renamedBinaryName))
-		if err != nil {
-			return err
-		}
+	renamedBinaryName, err := PromptRenameBinary(io, renameArgs)
+	if err != nil {
+		return err
+	}
+	err = copyFile(installation.DownloadedFilePath, filepath.Join(tmpExtractionPath, renamedBinaryName))
+	if err != nil {
+		return err
 	}
 	return nil
 }
 
-// InstallBinary will extract the binary and copy it to the ~/.stew/bin path
-func InstallBinary(downloadedFilePath string, repo string, systemInfo SystemInfo, lockFile *LockFile, overwriteFromUpgrade bool) (string, error) {
+// InstallBinary will extract the binary and copy it to the ~/.stew/bin path.
+func InstallBinary(rt config.Runtime, installation Installation, lockFile *LockFile, overwriteFromUpgrade bool) (string, error) {
+	tmpExtractionPath := rt.TmpPath
+	assetDownloadPath := rt.PkgPath
+	binaryInstallPath := rt.StewBinPath
 
-	tmpExtractionPath := systemInfo.StewTmpPath
-	assetDownloadPath := systemInfo.StewPkgPath
-	binaryInstallPath := systemInfo.StewBinPath
+	if err := os.RemoveAll(tmpExtractionPath); err != nil {
+		return "", errors.WithStack(err)
+	}
+	if err := os.MkdirAll(tmpExtractionPath, 0o755); err != nil {
+		return "", errors.WithStack(err)
+	}
 
-	err := extractBinary(downloadedFilePath, tmpExtractionPath)
+	err := extractBinary(rt, installation, tmpExtractionPath)
 	if err != nil {
 		return "", err
 	}
@@ -306,7 +309,7 @@ func InstallBinary(downloadedFilePath string, repo string, systemInfo SystemInfo
 		return "", err
 	}
 
-	binaryFile, binaryName, err := getBinary(allFilePaths, repo)
+	binaryFile, binaryName, err := getBinary(rt, allFilePaths, installation)
 	if err != nil {
 		return "", err
 	}
@@ -314,13 +317,17 @@ func InstallBinary(downloadedFilePath string, repo string, systemInfo SystemInfo
 	// Check if the binary already exists
 	for index, pkg := range lockFile.Packages {
 		previousAssetPath := filepath.Join(assetDownloadPath, pkg.Asset)
-		newAssetPath := downloadedFilePath
+		newAssetPath := installation.DownloadedFilePath
 		var overwrite bool
 		if pkg.Binary == binaryName {
-			if !overwriteFromUpgrade {
-				overwrite, err = WarningPromptConfirm(fmt.Sprintf("The binary %v version: %v is already installed, would you like to overwrite it?", constants.YellowColor(binaryName), constants.YellowColor(pkg.Tag)))
+			if !overwriteFromUpgrade && !installation.BatchMode {
+				overwrite, err = prompt.ConfirmWarn(rt,
+					fmt.Sprintf("The binary %v version: %v is already installed, "+
+						"would you like to overwrite it?",
+						constants.YellowColor(binaryName),
+						constants.YellowColor(pkg.Tag)))
 				if err != nil {
-					os.RemoveAll(newAssetPath)
+					_ = os.RemoveAll(newAssetPath)
 					return "", err
 				}
 			} else {
@@ -328,9 +335,9 @@ func InstallBinary(downloadedFilePath string, repo string, systemInfo SystemInfo
 			}
 
 			if overwrite {
-				err := os.RemoveAll(previousAssetPath)
+				err = os.RemoveAll(previousAssetPath)
 				if err != nil {
-					return "", err
+					return "", errors.WithStack(err)
 				}
 
 				if !overwriteFromUpgrade {
@@ -339,16 +346,15 @@ func InstallBinary(downloadedFilePath string, repo string, systemInfo SystemInfo
 						return "", err
 					}
 				}
-
 			} else {
 				err = os.RemoveAll(newAssetPath)
 				if err != nil {
-					return "", err
+					return "", errors.WithStack(err)
 				}
 
 				err = os.RemoveAll(tmpExtractionPath)
 				if err != nil {
-					return "", err
+					return "", errors.WithStack(err)
 				}
 
 				return "", AbortBinaryOverwriteError{Binary: pkg.Binary}
@@ -363,48 +369,26 @@ func InstallBinary(downloadedFilePath string, repo string, systemInfo SystemInfo
 
 	err = os.RemoveAll(tmpExtractionPath)
 	if err != nil {
-		return "", err
+		return "", errors.WithStack(err)
 	}
 
 	return binaryName, nil
 }
 
+// RenameBinaryArgs contains information about the renaming process.
+type RenameBinaryArgs struct {
+	Default   string
+	BatchMode bool
+}
+
 // PromptRenameBinary takes in the original name of the binary and will return the new name of the binary.
-func PromptRenameBinary(originalBinaryName string) (string, error) {
-	renamedBinaryName, err := warningPromptInput("Rename the binary?", originalBinaryName)
+func PromptRenameBinary(io terminal.Terminal, rename RenameBinaryArgs) (string, error) {
+	if rename.BatchMode {
+		return rename.Default, nil
+	}
+	renamedBinaryName, err := prompt.InputWarn(io, "Rename the binary?", rename.Default)
 	if err != nil {
 		return "", err
 	}
 	return renamedBinaryName, nil
-}
-
-// ResolvePath will resolve the full path for an input path
-func ResolvePath(filePath string) (string, error) {
-	var resolvedPath string
-	var err error
-	resolvedPath = filePath
-
-	resolvedPath = strings.ReplaceAll(resolvedPath, "\"", "")
-
-	if strings.HasPrefix(resolvedPath, "~") {
-		homeDir, err := os.UserHomeDir()
-		if err != nil {
-			return "", err
-		}
-
-		resolvedPath = filepath.Join(homeDir, strings.TrimLeft(resolvedPath, "~"))
-	}
-
-	resolvedPath = os.ExpandEnv(resolvedPath)
-
-	if !filepath.IsAbs(resolvedPath) {
-		resolvedPath, err = filepath.Abs(resolvedPath)
-		if err != nil {
-			return "", err
-		}
-	}
-
-	resolvedPath = strings.TrimRight(resolvedPath, "/")
-
-	return resolvedPath, nil
 }
